% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockCountingProcess.R
\name{block_counting_process}
\alias{block_counting_process}
\title{\code{block_counting_process} return a the block counting process for a given sample size
as a \code{mult_cont_phase_type} object.}
\usage{
block_counting_process(n)
}
\arguments{
\item{n}{Number of samples}
}
\value{
ph_rew_ob
A \code{mult_cont_phase_type} representation of the block counting process of size n
}
\description{
\code{block_counting_process} return a the block counting process for a given sample size
as a \code{mult_cont_phase_type} object.
}
\examples{
block_counting_process(8)

}
