% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phase_type.R
\name{check_phase_type}
\alias{check_phase_type}
\title{Generator Function for Phase-Type Distributions}
\usage{
check_phase_type(subint_mat, init_probs, reward_mat, round_zero)
}
\description{
Generator Function for Phase-Type Distributions
}
