% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phase_type.R
\name{check_reward}
\alias{check_reward}
\title{Checking reward matrix}
\usage{
check_reward(reward_mat, init_probs)
}
\arguments{
\item{reward_mat}{Reward matrix}

\item{init_probs}{Initial probability vector.}
}
\description{
This function checks whether the shape of the reward
matrix is correct.
}
\keyword{internal}
