% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_phase_type.R
\name{cont_phase_type}
\alias{cont_phase_type}
\title{\code{cont_phase_type} class}
\usage{
cont_phase_type(subint_mat = NULL, init_probs = NULL)
}
\arguments{
\item{subint_mat}{matrix}

\item{init_probs}{vector, a one-row matrix or \code{NULL}.}
}
\description{
Description of the class \code{cont_phase_type}, which represents continuous phase-type
distributions.
}
\details{
\code{cont_phase_type} is the generator function for the continuous phase-type distribution class
of the same name, which inherits from \code{list}. The class is generated by supplying a sub-intensity
matrix and an optional initial probability vector. If the initial probabilities are not specified,
then they are automatically generated as the first state having a probability of 1 and the rest
a probability of 0.
}
\examples{
subintensity_matrix = matrix(c(-1.5, 0, 0, 1.5, -1, 0, 0, 1, -0.5), ncol = 3)
cont_phase_type(subintensity_matrix)

#---

subintensity_matrix = matrix(c(-1.5, 0, 0, 1.5, -1, 0, 0, 1, -0.5), ncol = 3)
initial_probabilities = c(0.9, 0.1, 0)
cont_phase_type(subintensity_matrix, initial_probabilities)

}
