% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_reward.R
\name{cont_reward}
\alias{cont_reward}
\title{\code{cont_reward}}
\usage{
cont_reward(phase_type = NULL, init_probs = NULL,
subint_mat = NULL, reward_vec)
}
\arguments{
\item{phase_type}{a \code{cont_phase_type}
class object or \code{NULL} (default),}

\item{init_probs}{the initials probability of the Markov process (\eqn{\pi})
should be a vector, a one-row matrix or \code{NULL} (default).}

\item{subint_mat}{the subintensity matrix of the Markov process (\strong{T})
should be a square matrix or \code{NULL} (default).}

\item{reward}{the reward given to each state of the Markov process,
should be a vector or a one row matrix.}
}
\value{
Transformed \code{cont_phase_type} class object.
}
\description{
Transform a variable following a continuous
phase-type distribution according to a non-negative reward vector.
}
\details{
This function will perfom a reward transformation on a variable following
a continuous phase-type distribution. The reward should be non-negative.

Either \code{phase_type} or \code{init_probs}
and \code{subint_mat} should be filled.
if both are filled \code{phase_type} will be used.
If there is no init_probs (and no phase_type) the first state will have an
initial probability of 1.
}
\examples{

##======================================================
## Example 3.1.35 in the book of Bladt & Nielsen (2017)
##======================================================

subint_mat <- matrix(c(-3, 1, 1,
                        2, -2, 1,
                        0, 1, -2), nrow = 3)
init_probs <- c(0.5, 0.25, 0.25)
reward <- c(1, 0, 1)

cont_reward(init_probs = init_probs,
            subint_mat = subint_mat, reward = reward)

##===================================================
## Example
##===================================================

}
\references{
Bladt, M., & Nielsen, B. F. (2017).
Matrix-exponential distributions in applied probability (Vol. 81).
New York: Springer.
}
\seealso{
\code{\link{cont_phase_type}}
}
