% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convph.R
\name{convph}
\alias{convph}
\title{Convolution of Continuous Phase-Type Distributions}
\usage{
convph(ph1 = NULL, ph2 = NULL)
}
\arguments{
\item{ph1}{cont_phase_type class}

\item{ph2}{cont_phase_type class.}
}
\description{
\code{convph} is the function allowing convolution between two phase_type object
building a new phase_type object.
}
\examples{
ph1 = PH(matrix(c(-3, 0, 0, 1, -2, 0, 0, 1, -1), ncol = 3), c(0.25,0.25,0.5))
ph2 = PH(matrix(c(-1.5, 0, 0, 1.5, -1, 0, 0, 1, -0.5), ncol = 3), c(0.25,0.25,0.5))

convph(ph1,ph2)
#---

}
