% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc_phase_type.R
\name{disc_phase_type}
\alias{disc_phase_type}
\title{\code{disc_phase_type} class}
\usage{
disc_phase_type(subint_mat = NULL, init_probs = NULL)
}
\arguments{
\item{subint_mat}{matrix or \code{NULL} (default).}

\item{init_probs}{vector, a one-row matrix or \code{NULL} (default).}
}
\description{
Description of the class \code{disc_phase_type}, which represents discrete phase-type
distributions.
}
\details{
\code{disc_phase_type} is the generator function for the discrete phase-type distribution class
of the same name, which inherits from \code{list}. The class is generated by supplying a sub-intensity
matrix and an optional initial probability vector. If the initial probabilities are not specified,
then a they are automatically generated as the first state having a probability of 1 and the rest
a probability of 0.
}
\examples{
subintensity_matrix = matrix(c(0.4, 0, 0, 0.24, 0.4, 0, 0.12, 0.2, 0.5), ncol = 3)
disc_phase_type(subintensity_matrix)

#---

subintensity_matrix = matrix(c(0.4, 0, 0, 0.24, 0.4, 0, 0.12, 0.2, 0.5), ncol = 3)
initial_probabilities = c(0.9, 0.1, 0)
disc_phase_type(subintensity_matrix, initial_probabilities)

}
