% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc_reward.R
\name{disc_reward}
\alias{disc_reward}
\title{\code{disc_reward}}
\usage{
disc_reward(phase_type = NULL, init_probs = NULL,
subint_mat = NULL, reward_vec)
}
\value{
Transformed \code{disc_phase_type} class object.
}
\description{
Transform a variable following a discrete
phase-type distribution according to a non-negative reward vector.
}
\details{
This function will perfom a reward transformation on a variable following
a discrete phase-type distribution.
The reward should be non-negative integers.

Either \code{phase_type} or \code{init_probs}
and \code{subint_mat} should be filled.
if both are filled \code{phase_type} will be used.
If there is no init_probs (and no phase_type) the first state will have an
initial probability of 1.
}
\examples{

##========================================
##
##========================================

subint_mat <- matrix(c(0.4, 0, 0,
                      0.24, 0.4, 0,
                      0.12, 0.2, 0.5), ncol = 3)
init_probs <- c(0.9, 0.1, 0)
reward <- c(1,0,4)

disc_reward(init_probs = init_probs,
            subint_mat = subint_mat, reward = reward)
#---


}
\references{

}
\seealso{
\code{\link{disc_phase_type}}
}
