% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPH_construction.R
\name{DPHrep}
\alias{DPHrep}
\title{Construction of DPH-representation}
\usage{
DPHrep(bM, bA, ba)
}
\arguments{
\item{bM}{Subtransition probabilities un the underlying discrete Markov chain (cf. Figure 6).}

\item{bA}{Statespace of the underlying block-counting process}

\item{ba}{Vector of integer coeffcients}
}
\value{
List consisting of
bMt: The constructed matrix subtransition probabilities.
sizes_of_blocks: The sizes of the constructed blocks.
}
\description{
\code{DPHrep} computes the representation of of a integer-linear-combination of the Site Frequency as a discrete phase-type distribution.
The construction is described in Section 7.1 of Hobolth (2020).
}
\examples{
\dontrun{
ba = c(1,2,3)
ph_bcp = block_counting_process(4)
subintensity_matrix = ph_bcp$subint_mat
bA = ph_bcp$reward_mat
ph = PH(subintensity_matrix)
ph_rew_obj = reward_phase_type(ph, rowSums(rew_mat))
bS = ph_rew_obj$subint_mat
bM = solve(diag(dim(bS)[1])-(2/theta)*bS)
DPHrep(ba,bM)
}

}
