% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phtype_functions.R
\name{dphtype}
\alias{dphtype}
\alias{qphtype}
\alias{pphtype}
\alias{rphtype}
\title{Density, distribution function, quantile function and random generation for
phase-type distributions.}
\usage{
dphtype(x, obj)

pphtype(q, obj)

qphtype(p, obj)

rphtype(n, obj)

qphtype(p, obj)

pphtype(q, obj)

rphtype(n, obj)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{obj}{an object of class \code{cont_phase_type} or
\code{disc_phase_type}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If length(n) > 1, the length is taken to be
the number required.}
}
\value{
\code{dphtype} gives the density, \code{pphtype} gives the
distribution function, \code{qphtype} gives the quantile function,
and \code{rphtype} generates random deviates.

The length of the result is determined by \code{n} for \code{rphtype},
and is the maximum of the lengths of the numerical arguments for the other
functions.

The numerical arguments other than \code{n} are recycled to the length of the
result. Only the first elements of the logical arguments are used.
}
\description{
Density, distribution function, quantile function and random generation for
phase-type distributions.
}
\section{Functions}{
\itemize{
\item \code{dphtype}: Density function.

\item \code{qphtype}: Quantile Function

\item \code{pphtype}: Distribution function.

\item \code{rphtype}: Random number generator
}}

\seealso{
\link[stats]{Distributions}
}
