% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDPH.R
\name{MDPH}
\alias{MDPH}
\title{Multivariate discrete phase-type distribution}
\usage{
MDPH(
  subint_mat = NULL,
  init_probs = NULL,
  reward_mat = NULL,
  round_zero = NULL
)
}
\description{
Multivariate discrete phase-type distribution
}
