% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{moment_mph}
\alias{moment_mph}
\title{Moments of the multivariate continuous phase-type distribution}
\usage{
moment_mph(obj, v)
}
\arguments{
\item{obj}{a mult_cont_phase_type.}

\item{v}{a vector or an integer.}
}
\description{
This function calculates the moments for the multivariate phase-type
distributions.
}
\details{
The variables for which the moments are calculated can be specified in
the \code{v} vector as indices in the reward matrix of the
\code{mult_cont_phase_type} object.
}
