% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{moment_ph}
\alias{moment_ph}
\title{Moments of the univariate continuous phase-type distribution}
\usage{
moment_ph(obj, m)
}
\arguments{
\item{obj}{a mult_cont_phase_type.}

\item{m}{an integer.}
}
\description{
This function calculates the moments for the phase-type
distributions.
}
\details{
The order of the moment can be specified by \code{m} for a
\code{cont_phase_type} object.
}
\keyword{internal}
