% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPH.R
\name{MPH}
\alias{MPH}
\title{Multivariate continuous phase-type distribution}
\usage{
MPH(subint_mat = NULL, init_probs = NULL, reward_mat = NULL, round_zero = NULL)
}
\description{
Multivariate continuous phase-type distribution
}
