% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_phase_type.R
\name{mult_phase_type}
\alias{mult_phase_type}
\title{\code{mult_phase_type} class}
\usage{
mult_phase_type(subint_mat = NULL, reward_mat = NULL, init_probs = NULL)
}
\arguments{
\item{subint_mat}{matrix}

\item{reward_mat}{matrix}

\item{init_probs}{vector, a one-row matrix or \code{NULL}.}
}
\description{
Description of the class \code{mult_phase_type}, which represents the multivariate phase-type
distribution.
}
\details{
\code{cont_phase_type} is the generator function for the multivariate phase-type distribution class
of the same name, which inherits from \code{list}. The class is generated by supplying a sub-intensity
matrix, an optional initial probability vector and a reward matrix. If the initial probabilities are not specified,
then they are automatically generated as the first state having a probability of 1 and the rest
a probability of 0.
}
\examples{
subint_mat = matrix(c(-3, 0, 0, 2, -2, 0, 0, 1, -1), nrow = 3, ncol = 3)
reward_mat = matrix(sample(seq(0, 10), 6), nrow = 3, ncol = 2)
mult_phase_type(subint_mat, reward_mat)

}
