% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{perm}
\alias{perm}
\title{Permutations}
\usage{
perm(v)
}
\arguments{
\item{v}{a numeric vector.}
}
\value{
A matrix with all possible permutations.
}
\description{
This function calculates all possible permutations given a
numeric vector.
}
\keyword{internal}
