% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PH.R, R/check_phase_type.R
\name{PH}
\alias{PH}
\alias{summary.cont_phase_type}
\title{Continuous phase-type distribution}
\usage{
PH(subint_mat = NULL, init_probs = NULL, round_zero = NULL)

\method{summary}{cont_phase_type}(object, ...)
}
\arguments{
\item{object}{a phase-type object}

\item{...}{other arguments passed to methods}
}
\description{
Continuous phase-type distribution
}
\section{Methods (by generic)}{
\itemize{
\item \code{summary}: pretty summary of the class.
}}

