% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phtype_functions.R
\name{Phase-type}
\alias{Phase-type}
\alias{dphtype}
\alias{qphtype}
\alias{pphtype}
\alias{rphtype}
\title{The Phase-Type Distribution}
\usage{
dphtype(x, obj)

qphtype(p, obj)

pphtype(q, obj)

rphtype(n, obj)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{obj}{an object of class \code{cont_phase_type}, \code{disc_phase_type}
, \code{mult_cont_phase_type} or \code{mult_disc_phase_type}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If length(n) > 1, the length is taken to be
the number required.}
}
\value{
\code{dphtype} gives the density, \code{pphtype} gives the
distribution function, \code{qphtype} gives the quantile function,
and \code{rphtype} generates random deviates.

The length of the result is determined by \code{n} for \code{rphtype},
and is the maximum of the lengths of the numerical arguments for the other
functions.

The numerical arguments other than \code{n} are recycled to the length of the
result. Only the first elements of the logical arguments are used.
}
\description{
The Phase-Type Distribution
}
\details{
If the object provided is multivariate, each row of the result will
corresponds to each univariate reward transformation.
For \code{dphtype}, \code{qphtype} and \code{pphtype}, the inputs \code{x},
\code{p} and \code{q} can be matrices where in row i the i_th reward
transformation and in col j the j_th value of \code{x}, \code{p} or \code{q}
tested.
}
\section{Functions}{
\itemize{
\item \code{dphtype}: Density function.

\item \code{qphtype}: Quantile Function

\item \code{pphtype}: Distribution function.

\item \code{rphtype}: Random number generator
}}

\examples{

cont_phase_type <- matrix(c(-3, 0, 1,
                            2, -3, 1,
                            1, 1, -2), ncol = 3)
Y <- PH(cont_phase_type)
x <- seq(0, 2, length.out = 20)
plot(x, dphtype(x, Y), ylab = 'pdf')

# and for the discrete phase-type distribution

disc_phase_type <- matrix(c(0.4, 0, 0.2,
                            0.5, 0.3, 0.2,
                            0, 0.7, 0.2), ncol = 3)
Y <- DPH(disc_phase_type)
x <- seq(0, 20, 1) # Only integer for the x values.
plot(x, dphtype(x, Y), ylab = 'pmf')

rphtype(20, Y)

}
\seealso{
\link[stats]{Distributions} for other standard distributions.
}
