% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reward_disc.R
\name{reward_DPH}
\alias{reward_DPH}
\title{\code{disc_phase_type} class}
\usage{
reward_DPH(DPH = NULL, init_probs = NULL, subint_mat = NULL,
reward_vec = NULL)
}
\arguments{
\item{DPH}{a disc_phase_type object or \code{NULL} (default).}

\item{init_probs}{vector, a one-row matrix or \code{NULL} (default).}

\item{subint_mat}{matrix or \code{NULL} (default).}

\item{reward}{vector, a one-row matrix or \code{NULL} (default).}
}
\description{
Description of the class \code{disc_phase_type}, which represents discrete phase-type
distributions.
}
\details{
\code{reward_DPH} is the function to attribute a non-negative
reward vector to a discrete phase-type distribution.
}
\examples{

subint_mat = matrix(c(0.4, 0, 0, 0.24, 0.4, 0, 0.12, 0.2, 0.5), ncol = 3)
init_probs = c(0.9, 0.1, 0)
R = c(1,0,4)

reward_DPH(init_probs = init_probs, subint_mat = subint_mat, reward_vec = R)
------------------------



}
