% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul - Copie.R, R/simul.R
\name{sim_phase_type}
\alias{sim_phase_type}
\title{Phase-type simulations}
\usage{
sim_phase_type <- function(obj)

sim_phase_type <- function(obj)
}
\arguments{
\item{obj}{an object of class \code{disc_phase_type} or
\code{cont_phase_type}}
}
\description{
sim_phase_type implements a naive simulation scheme for simulation of
phase-type distributions.

sim_phase_type implements a naive simulation scheme for simulation of
phase-type distributions.
}
\examples{

##===========================##
## For continuous univariate ##
##===========================##

set.seed(2)
subintensity_matrix <- matrix(c(-1.5, 0, 0.5,
                               1.5, -1, 0,
                               0, 1, -1), ncol = 3)
obj <- phase_type(subintensity_matrix)

sim_phase_type(obj)

##===========================##
## For discrete multivariate ##
##===========================##

set.seed(2)
subintensity_matrix <- matrix(c(0.4, 0, 0.1,
                               0.24, 0.6, 0.3,
                               0.26, 0.4, 0.5), ncol = 3)
obj <- phase_type(subintensity_matrix)

sim_phase_type(obj)

##=============================##
## For continuous multivariate ##
##=============================##

subintensity_matrix <- matrix(c(-3, 0, 0,
                               2, -2, 0,
                               0, 1, -1), nrow = 3, ncol = 3)
reward_matrix = matrix(sample(seq(0, 10, 0.1), 6), nrow = 3, ncol = 2)
obj <- phase_type(subintensity_matrix, reward_mat = reward_matrix)
sim_phase_type(obj)

##===========================##
## For discrete multivariate ##
##===========================##

subintensity_matrix <- matrix(c(0.4, 0, 0,
                               0.24, 0.4, 0,
                               0.12, 0.2, 0.5), ncol = 3)
reward_matrix <- matrix(sample(seq(0, 10), 6), nrow = 3, ncol = 2)
phase_type(subintensity_matrix, reward_mat = reward_matrix)


##===========================##
## For continuous univariate ##
##===========================##

set.seed(2)
subintensity_matrix <- matrix(c(-1.5, 0, 0.5,
                               1.5, -1, 0,
                               0, 1, -1), ncol = 3)
obj <- phase_type(subintensity_matrix)

sim_phase_type(obj)

##===========================##
## For discrete multivariate ##
##===========================##

set.seed(2)
subintensity_matrix <- matrix(c(0.4, 0, 0.1,
                               0.24, 0.6, 0.3,
                               0.26, 0.4, 0.5), ncol = 3)
obj <- phase_type(subintensity_matrix)

sim_phase_type(obj)

##=============================##
## For continuous multivariate ##
##=============================##

subintensity_matrix <- matrix(c(-3, 0, 0,
                               2, -2, 0,
                               0, 1, -1), nrow = 3, ncol = 3)
reward_matrix = matrix(sample(seq(0, 10, 0.1), 6), nrow = 3, ncol = 2)
phase_type(subintensity_matrix, reward_mat = reward_matrix)

##===========================##
## For discrete multivariate ##
##===========================##

subintensity_matrix <- matrix(c(0.4, 0, 0,
                               0.24, 0.4, 0,
                               0.12, 0.2, 0.5), ncol = 3)
reward_matrix <- matrix(sample(seq(0, 10), 6), nrow = 3, ncol = 2)
phase_type(subintensity_matrix, reward_mat = reward_matrix)

}
