% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{sim_rew_phase_type}
\alias{sim_rew_phase_type}
\title{sim_rew_phase_type implements a naive simulation scheme for simulation of
multivariate phase-type distributions.}
\usage{
sim_rew_phase_type(R, obj, reward_mat = NULL)
}
\arguments{
\item{R}{number of replications}

\item{obj}{an object of class \code{mult_cont_phase_type} or
\code{cont_phase_type}}

\item{reward_mat}{A p times n matrix of rewards. If a vector is
provided it is converted to a matrix and univariate rewards are assumed
the number of columns of reward_mat must match subint_mat}
}
\value{
An n times R vector
}
\description{
sim_rew_phase_type implements a naive simulation scheme for simulation of
multivariate phase-type distributions.
}
