% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApproximateCharFun.R
\name{simple_smoothe}
\alias{simple_smoothe}
\title{Simple smoothing}
\usage{
simple_smoothe(x, svec = c(0.1, 0.2, 0.4, 0.2, 0.1))
}
\arguments{
\item{x}{input to be smoothed}

\item{svec}{smoothing vector}
}
\description{
\code{simple_smoothe} computes a simple moving weighted average of a input vector \code{x}. The weight vector must have an odd number of entries, and defaults to
\code{c(0.1,0.20,0.4,0.20,0.1)}
}
\examples{
\dontrun{
smoothed_yvals = simple_smoothe(yvals)
smoothed_yvals = simple_smoothe(yvals,c(0.2,0.6,0.2))
}

}
