% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.cont_phase_type}
\alias{summary.cont_phase_type}
\title{Pretty summary of the \code{cont_phase_type} class.}
\usage{
\method{summary}{cont_phase_type}(object, ...)
}
\arguments{
\item{object}{a cont_phase_type object}

\item{...}{other arguments passed to methods}
}
\value{
This function prints a nicely-formatted summary of
a \code{cont_phase_type} object. The summary includes the
sub-intensity matrix, the initial probabilities, the defect,
the mean and the variance of the phase-type object.
}
\description{
Pretty summary of the \code{cont_phase_type} class.
}
\examples{

ph <- PH(matrix(c(-3, 0, 1,
               2, -3, 1,
               1, 1, -2), ncol = 3))

summary(ph)

}
