% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_phase_type.R
\name{var}
\alias{var}
\title{Variance of cont_phase_type distributions}
\usage{
var(obj)
}
\arguments{
\item{obj}{a cont_phase_type or disc_phase_type object.}
}
\description{
Calculates the variance of continuous and discrete phase-type distributions,
represented by the \code{cont_phase_type} and the \code{disc_phase_type} classes
respectively
}
