% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phtype_functions.R
\name{dphtype}
\alias{dphtype}
\alias{qphtype}
\alias{pphtype}
\alias{rphtype}
\title{Density, distribution function, quantile function and random generation for
phase-type distributions}
\usage{
dphtype(x, obj)

qphtype(p, obj)

pphtype(q, obj)

rphtype(n, obj)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{obj}{an object of class \code{cont_phase_type} or \code{disc_phase_type}.}

\item{p}{vector of probabilities.}

\item{n}{number of samples to draw.}
}
\description{
Density, distribution function, quantile function and random generation for
phase-type distributions
}
\section{Functions}{
\itemize{
\item \code{dphtype}: Density function.

\item \code{qphtype}: Quantile Function

\item \code{pphtype}: Distribution function.

\item \code{rphtype}: Random number generator
}}

