% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mean}
\alias{mean}
\alias{mean.cont_phase_type}
\alias{mean.disc_phase_type}
\alias{mean.mult_cont_phase_type}
\title{Mean of phase-type distributions}
\usage{
mean(x, ...)

\method{mean}{cont_phase_type}(x, ...)

\method{mean}{disc_phase_type}(x, ...)

mean(x, v = NULL, ...)
}
\arguments{
\item{x}{a \code{cont_phase_type}, \code{disc_phase_type},
\code{mult_cont_phase_type} or \code{mult_disc_phase_type} object}

\item{...}{other arguments passed to methods}

\item{v}{NULL, integer or vector.}
}
\description{
Calculates the mean of continuous, discrete and multivariate phase-type
distributions, represented by the \code{cont_phase_type},
\code{disc_phase_type} and \code{mult_cont_phase_type} classes.
}
\details{
For the univariate case (\code{cont_phase_type} and \code{disc_phase_type}),
the mean of the distribution is returned.

In the case of multivariate phase-type distributions three
different usages can be distinguished:
\itemize{
 \item{If \code{v = NULL} (default), the means of all the variables defined by
 the sub-intensity matrix are returned}
 \item{If \code{v} is an integer, then the mean of the variable with the specified index in the
 reward matrix is returned.}
 \item{If \code{v} is a vector, then the means of the variables defined
by those indices will be returned.}
}
}
