% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{var}
\alias{var}
\alias{var.cont_phase_type}
\alias{var.disc_phase_type}
\alias{var.mult_cont_phase_type}
\alias{var.mult_disc_phase_type}
\title{Variance of phase-type distributions}
\usage{
var(obj, ...)

\method{var}{cont_phase_type}(obj, ...)

\method{var}{disc_phase_type}(obj, ...)

var(obj, v = NULL, ...)

var(obj, v = NULL, ...)
}
\arguments{
\item{obj}{a \code{cont_phase_type}, \code{disc_phase_type},
\code{mult_cont_phase_type} or \code{mult_disc_phase_type} object}

\item{...}{other arguments passed to methods}

\item{v}{NULL, integer or vector of length 2.}
}
\description{
Calculates the variance of continuous, discrete and multivariate phase-type
distributions, represented by the \code{cont_phase_type},
\code{disc_phase_type} and \code{mult_cont_phase_type} classes respectively.
}
\details{
For  the univariate case (\code{cont_phase_type} and \code{disc_phase_type}),
the variance of the distribution is returned.

In the case of multivariate phase-type distributions three
different usages can be distinguished:
\itemize{
 \item{If \code{v = NULL} (default), then a variance-covariance matrix of all
 the variables specified in the reward matrix are returned, where variances
 are in the diagonal and covariances in the rest of the matrix element.}
 \item{If \code{v} is an integer, then the variance of the variable encoded
 by the \code{v} index in the reward matrix is returned.}
 \item{If \code{v} is a vector of length 2, then the covariance between the
 two variables encoded by the \code{v} indices in the reward matrix is
 returned.}
}
}
